Fix file sweeper for missed files
When disconnected for some period of time, files can build up.  The
sweeper uploads these files.  It includes a check of the timestamp of
the file, which it gets by parsing the filename.  In order for this
check to work correctly, the filename suffix of ".json" must be removed
from the filename.