Ensure only one sweeper is running at once
setInterval() can lead to concurrent executions if the work it
scheduling takes longer than the interval itself.  While our timer is
set to 10min, so this is unlikely, the sweeper is designed to run to
clear a large backlog of work after a network outage.  Safer to use the
recursive setTimeout() pattern recommended by MDN.